Add gravitational_trace.nocrypto build tag
