+++ .github/workflows/test.yaml
@@ -20,3 +20,5 @@ jobs:
           go-version-file: go.mod
       - name: Test
         run: go test -race ./...
+      - name: Test (nocrypto)
+        run: go test -race -tags gravitational_trace.nocrypto ./...
