+++ errors_crypto.go
@@ -1,3 +1,5 @@
+//go:build !gravitational_trace.nocrypto
+
 // Copyright 2026 Gravitational, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,31 +14,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal
+package trace
 
-// TraverseErr traverses the err error chain until fn returns true.
-// Traversal stops on nil errors, fn(nil) is never called.
-// Returns true if fn matched, false otherwise.
-func TraverseErr(err error, fn func(error) (ok bool)) (ok bool) {
-	if err == nil {
-		return false
-	}
+import "crypto/x509"
 
-	if fn(err) {
-		return true
-	}
-
-	switch err := err.(type) {
-	case interface{ Unwrap() error }:
-		return TraverseErr(err.Unwrap(), fn)
-
-	case interface{ Unwrap() []error }:
-		for _, err2 := range err.Unwrap() {
-			if TraverseErr(err2, fn) {
-				return true
-			}
-		}
-	}
-
-	return false
-}
+// these type aliases are used by [ConvertSystemError]; builds with the nocrypto
+// tag will define them differently
+type (
+	x509SystemRootsError      = x509.SystemRootsError
+	x509UnknownAuthorityError = x509.UnknownAuthorityError
+)
